Moves supplier related classes according to project restured
