+++ app/src/main/java/com/pragati/karuna/request/adapter/SupplierAdapter.kt
@@ -1,4 +1,4 @@
-package com.pragati.karuna.ui.suppliers
+package com.pragati.karuna.request.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,13 +6,15 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.pragati.karuna.R
-import com.pragati.karuna.viewmodel.SuppliersViewModel
+import com.pragati.karuna.request.viewmodel.SuppliersViewModel
 
 class SupplierAdapter(private val viewModel: SuppliersViewModel) : RecyclerView.Adapter<SupplierAdapter.SupplierCell>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SupplierCell {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.supplier_details_cell, parent, false)
-        return SupplierCell(view)
+        return SupplierCell(
+            view
+        )
     }
 
     override fun getItemCount(): Int {
