+++ app/src/main/res/layout/fragment_add_suppliers.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_marginStart="16dp"
     android:layout_marginEnd="16dp"
-    tools:context=".ui.suppliers.AddSuppliersFragment">
+    tools:context=".request.ui.AddSuppliersFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/suppliers_list_container"
